Move python leveling in optics_specific_tools.py
