+++ python_examples/hl_lhc_collisions_python/optics_specific_tools.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 import pymask as pm
 
 # The parts marked by (*) in following need to be
@@ -53,19 +55,6 @@ def build_sequence(mad, beam):
 def apply_optics(mad, optics_file):
     mad.call(optics_file)
 
-def check_beta_at_ips_against_madvars(beam, twiss_df, variable_dicts, tol):
-    twiss_value_checks=[]
-    for iip, ip in enumerate([1,2,5,8]):
-        for plane in ['x', 'y']:
-            # (*) Adapet based on knob definitions
-            twiss_value_checks.append({
-                    'element_name': f'ip{ip}:1',
-                    'keyword': f'bet{plane}',
-                    'varname': f'bet{plane}ip{ip}b{beam}',
-                    'tol': tol[iip]})
-
-    pm.check_twiss_against_madvars(twiss_value_checks, twiss_df, variable_dicts)
-
 
 def set_optics_specific_knobs(mad, knob_settings, mode=None):
 
@@ -80,22 +69,6 @@ def set_optics_specific_knobs(mad, knob_settings, mode=None):
     mad.input('on_alice := on_alice_normalized * 7000./nrj;')
     mad.input('on_lhcb := on_lhcb_normalized * 7000./nrj;')
 
-def check_separations_at_ips_against_madvars(twiss_df_b1, twiss_df_b2,
-        variables_dict, tol):
-
-    separations_to_check = []
-    for iip, ip in enumerate([1,2,5,8]):
-        for plane in ['x', 'y']:
-            # (*) Adapet based on knob definitions
-            separations_to_check.append({
-                    'element_name': f'ip{ip}:1',
-                    'scale_factor': -2*1e-3,
-                    'plane': plane,
-                    # knobs like on_sep1h, onsep8v etc
-                    'varname': f'on_sep{ip}'+{'x':'h', 'y':'v'}[plane],
-                    'tol': tol[iip]})
-    pm.check_separations_against_madvars(separations_to_check,
-            twiss_df_b1, twiss_df_b2, variables_dict)
 
 def twiss_and_check(mad, sequences_to_check, twiss_fname,
         tol_beta=1e-3, tol_sep=1e-6, save_twiss_files=True,
@@ -121,7 +94,7 @@ def twiss_and_check(mad, sequences_to_check, twiss_fname,
     if check_betas_at_ips:
         for ss in sequences_to_check:
             tt = twiss_dfs[ss]
-            check_beta_at_ips_against_madvars(beam=ss[-1],
+            _check_beta_at_ips_against_madvars(beam=ss[-1],
                     twiss_df=tt,
                     variable_dicts=var_dict,
                     tol=tol_beta)
@@ -130,7 +103,7 @@ def twiss_and_check(mad, sequences_to_check, twiss_fname,
     if check_separations_at_ips:
         twiss_df_b1 = twiss_dfs['lhcb1']
         twiss_df_b2 = twiss_dfs['lhcb2']
-        check_separations_at_ips_against_madvars(twiss_df_b1, twiss_df_b2,
+        _check_separations_at_ips_against_madvars(twiss_df_b1, twiss_df_b2,
                 var_dict, tol=tol_sep)
         print('IP separation test against knobs passed!')
 
@@ -139,3 +112,63 @@ def twiss_and_check(mad, sequences_to_check, twiss_fname,
     other_data['summ_dfs'] = summ_dfs
 
     return twiss_dfs, other_data
+
+def lumi_control(mad, python_parameters, mask_parameters, knob_names):
+    from scipy.optimize import least_squares
+
+    # Leveling in IP8
+    sep_plane_ip8 = python_parameters['sep_plane_ip8']
+    sep_knobname_ip8 = knob_names['sepknob_ip8_mm']
+
+    L_target_ip8 = mask_parameters['par_lumi_ip8']
+    def function_to_minimize_ip8(sep8_m):
+        my_dict_IP8=lumi.get_luminosity_dict(
+            mad, twiss_dfs, 'ip8', mask_parameters['par_nco_IP8'])
+        my_dict_IP8[sep_plane_ip8 + '_1']=np.abs(sep8_m)
+        my_dict_IP8[sep_plane_ip8 + '_2']=-np.abs(sep8_m)
+        return np.abs(lumi.L(**my_dict_IP8) - L_target_ip8)
+    sigma_sep_b1_ip8=np.sqrt(twiss_dfs['lhcb1'].loc['ip8:1']['bet'+sep_plane_ip8]
+               * mad.sequence.lhcb1.beam['e'+sep_plane_ip8])
+    optres_ip8=least_squares(function_to_minimize_ip8, sigma_sep_b1_ip8)
+    mad.globals[sep_knobname_ip8] = (np.sign(mad.globals[sep_knobname_ip8])
+                                * np.abs(optres_ip8['x'][0])*1e3)
+
+    # Halo collision in IP2
+    sep_plane_ip2 = python_parameters['sep_plane_ip2']
+    sep_knobname_ip2 = knob_names['sepknob_ip2_mm']
+    sigma_sep_b1_ip2=np.sqrt(
+            twiss_dfs['lhcb1'].loc['ip2:1']['bet'+sep_plane_ip2]
+            * mad.sequence.lhcb1.beam['e'+sep_plane_ip2])
+    mad.globals[sep_knobname_ip2] = (np.sign(mad.globals[sep_knobname_ip2])
+            * mask_parameters['par_fullsep_in_sigmas_ip2']*sigma_sep_b1_ip2/2*1e3)
+
+
+def _check_beta_at_ips_against_madvars(beam, twiss_df, variable_dicts, tol):
+    twiss_value_checks=[]
+    for iip, ip in enumerate([1,2,5,8]):
+        for plane in ['x', 'y']:
+            # (*) Adapet based on knob definitions
+            twiss_value_checks.append({
+                    'element_name': f'ip{ip}:1',
+                    'keyword': f'bet{plane}',
+                    'varname': f'bet{plane}ip{ip}b{beam}',
+                    'tol': tol[iip]})
+
+    pm.check_twiss_against_madvars(twiss_value_checks, twiss_df, variable_dicts)
+
+def _check_separations_at_ips_against_madvars(twiss_df_b1, twiss_df_b2,
+        variables_dict, tol):
+
+    separations_to_check = []
+    for iip, ip in enumerate([1,2,5,8]):
+        for plane in ['x', 'y']:
+            # (*) Adapet based on knob definitions
+            separations_to_check.append({
+                    'element_name': f'ip{ip}:1',
+                    'scale_factor': -2*1e-3,
+                    'plane': plane,
+                    # knobs like on_sep1h, onsep8v etc
+                    'varname': f'on_sep{ip}'+{'x':'h', 'y':'v'}[plane],
+                    'tol': tol[iip]})
+    pm.check_separations_against_madvars(separations_to_check,
+            twiss_df_b1, twiss_df_b2, variables_dict)
